convert archive package to storage package. move most things to storage package
